requirements.txt
pillow==11.0.0
tk==0.1.0
matplotlib==3.9.2
minizinc==0.9.0